programsonline_backend/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.5
django-braces==1.14.0
django-cors-headers==3.6.0
django-environ==0.4.5
django-filebrowser-no-grappelli==3.8.0
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-sslserver==0.22
django-tinymce4-lite==1.8.0
django-uuslug==1.2.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jsmin==2.2.2
kombu==5.0.2
oauthlib==3.1.0
Pillow==8.1.0
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
